refactor: 아카데미 조회 응답 내 activities 필드명을 academies로 변경
