+++ src/main/java/kr/tennispark/activity/user/presentation/dto/response/GetAcademyResponse.java
@@ -3,7 +3,7 @@ package kr.tennispark.activity.user.presentation.dto.response;
 import java.util.List;
 import kr.tennispark.activity.common.domain.Activity;
 
-public record GetAcademyResponse(List<AcademyDTO> activities) {
+public record GetAcademyResponse(List<AcademyDTO> academies) {
 
     public static GetAcademyResponse of(List<Activity> activities) {
         List<AcademyDTO> dtos = activities.stream()
